+++ kubernetes-client.sln
@@ -43,16 +43,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkipTestLogger", "tests\Ski
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "customResource", "examples\customResource\customResource.csproj", "{95672061-5799-4454-ACDB-D6D330DB1EC4}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "util", "util", "{B9F72EFC-551A-4A0C-B4DC-DBE5AF8F5FE8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubernetesClient.Util", "util\src\KubernetesClient.Util\KubernetesClient.Util.csproj", "{45B6236C-C57A-4622-A631-8AD3AA1DB768}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{96C503A9-4DF5-43FF-A80E-7E96F7C35CA8}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{AEE9B8E6-7D40-46B2-A857-720CF03C47DA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubernetesClient.Util.Tests", "util\tests\KubernetesClient.Util.Tests\KubernetesClient.Util.Tests.csproj", "{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -255,30 +245,6 @@ Global
 		{95672061-5799-4454-ACDB-D6D330DB1EC4}.Release|x64.Build.0 = Release|Any CPU
 		{95672061-5799-4454-ACDB-D6D330DB1EC4}.Release|x86.ActiveCfg = Release|Any CPU
 		{95672061-5799-4454-ACDB-D6D330DB1EC4}.Release|x86.Build.0 = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|x64.Build.0 = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Debug|x86.Build.0 = Debug|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|Any CPU.Build.0 = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|x64.ActiveCfg = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|x64.Build.0 = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|x86.ActiveCfg = Release|Any CPU
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768}.Release|x86.Build.0 = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|x64.Build.0 = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Debug|x86.Build.0 = Debug|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|x64.ActiveCfg = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|x64.Build.0 = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|x86.ActiveCfg = Release|Any CPU
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -300,10 +266,6 @@ Global
 		{5056C4A2-5E12-4C16-8DA7-8835DA58BFF2} = {8AF4A5C2-F0CE-47D5-A4C5-FE4AB83CA509}
 		{4D2AE427-F856-49E5-B61D-EA6B17D89051} = {8AF4A5C2-F0CE-47D5-A4C5-FE4AB83CA509}
 		{95672061-5799-4454-ACDB-D6D330DB1EC4} = {B70AFB57-57C9-46DC-84BE-11B7DDD34B40}
-		{96C503A9-4DF5-43FF-A80E-7E96F7C35CA8} = {B9F72EFC-551A-4A0C-B4DC-DBE5AF8F5FE8}
-		{45B6236C-C57A-4622-A631-8AD3AA1DB768} = {96C503A9-4DF5-43FF-A80E-7E96F7C35CA8}
-		{AEE9B8E6-7D40-46B2-A857-720CF03C47DA} = {B9F72EFC-551A-4A0C-B4DC-DBE5AF8F5FE8}
-		{9B55CB07-0FA9-4DB6-9388-36A1C3B5B319} = {AEE9B8E6-7D40-46B2-A857-720CF03C47DA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {049A763A-C891-4E8D-80CF-89DD3E22ADC7}
