+++ src/KubernetesClient/Util/Informer/Cache/DeltaType.cs
@@ -1,4 +1,4 @@
-namespace k8s.Util.Cache
+namespace k8s.Util.Informer.Cache
 {
     public enum DeltaType
     {
