+++ src/KubernetesClient/Util/Informer/Cache/IIndexer.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using k8s.Models;
 
-namespace k8s.Util.Cache
+namespace k8s.Util.Informer.Cache
 {
     public interface IIndexer<TApiType> : IStore<TApiType>
       where TApiType : class, IKubernetesObject<V1ObjectMeta>
