+++ src/KubernetesClient/Util/Informer/Cache/MutablePair.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace k8s.Util.Cache
+namespace k8s.Util.Informer.Cache
 {
     public class MutablePair<TLeft, TRight>
     {
