+++ tests/KubernetesClient.Tests/KubernetesClient.Tests.csproj
@@ -8,6 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+      <PackageReference Include="MartinCostello.Logging.XUnit" Version="0.1.2" />
       <PackageReference Include="FluentAssertions" Version="5.10.3" />
       <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
       <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="13.2.33" />
@@ -40,5 +41,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\KubernetesClient\KubernetesClient.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Util\Common" />
   </ItemGroup>
 </Project>
