+++ tests/KubernetesClient.Tests/Util/Informer/Cache/ListerTest.cs
@@ -2,9 +2,9 @@ using System.Linq;
 using FluentAssertions;
 using k8s.Models;
 using Xunit;
-using k8s.Util.Cache;
+using k8s.Util.Informer.Cache;
 
-namespace k8s.Util.Tests.Cache
+namespace k8s.Tests.Util.Informer.Cache
 {
     public class ListerTest
     {
