+++ util/src/KubernetesClient.Util/KubernetesClient.Util.csproj
@@ -35,4 +35,8 @@
         <PackageReference Include="Microsoft.Extensions.Logging" Version="2.2.0" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Utils" />
+    </ItemGroup>
+
 </Project>
